+++ packages/nextjs/src/components/_editor-only/global-warnings/warnings/form-numbers/FormNumbersWarning.tsx
@@ -13,7 +13,7 @@ export const FormNumbersWarning = ({ content }: { content: ContentProps }) => {
     };
 
     return hasInvalidFormNumbers() ? (
-        <li>
+        <li key="form-numbers-warning">
             Skjemanummer må være på formatet {'NAV XX-XX.XX'} eller {'NAV XX-XX.XXy'}
         </li>
     ) : null;
